1.1

+++ src/main/java/io/github/moulberry/alaawiicoinslost/Config.java
@@ -3,11 +3,11 @@ package io.github.moulberry.alaawiicoinslost;
 public class Config {
 
     public Position displayPosition = new Position(10, 10);
-    public int coinsLost = 0;
+    public float coinsLost = 0;
     public float coinLostRatio = 100;
     public boolean counterActive = true;
     public boolean hidden = false;
     public int totalDamageTaken = 0;
-    public int totalCoinsLost = 0;
+    public float totalCoinsLost = 0;
 
 }
